add support for big decimals

+++ src/main/java/com/antonkesy/jcalculator/tokenizer/token/value/ValueToken.java
@@ -2,6 +2,8 @@ package com.antonkesy.jcalculator.tokenizer.token.value;
 
 import com.antonkesy.jcalculator.tokenizer.token.Token;
 
+import java.math.BigDecimal;
+
 public interface ValueToken extends Token {
-    int getValue();
+    BigDecimal getValue();
 }
